added floors

+++ src/render/raycast.c
@@ -6,7 +6,7 @@
 /*   By: ikulik <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/19 15:30:07 by ikulik            #+#    #+#             */
-/*   Updated: 2025/08/30 13:39:57 by ikulik           ###   ########.fr       */
+/*   Updated: 2025/09/04 16:32:01 by ikulik           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,9 +49,9 @@ void	calculate_steps(t_game *data, t_ray *ray, float column)
 	ray->step_y.x = 0;
 	ray->step_y.y = 0;
 	if (fabs(ray->view.x - 0.0f) > __FLT_EPSILON__)
-		ray->step_x = mult_scalar(ray->view, 1.0f / fabs(ray->view.x));
+		ray->step_x = mult_scalar(ray->view, 1.0f / fabsf(ray->view.x));
 	if (fabs(ray->view.y - 0.0f) > __FLT_EPSILON__)
-		ray->step_y = mult_scalar(ray->view, 1.0f / fabs(ray->view.y));
+		ray->step_y = mult_scalar(ray->view, 1.0f / fabsf(ray->view.y));
 }
 
 char	get_hit_type(t_ray *ray, t_pos point)
